Added more models, controllers, and databases.
I have added the models for Classes, assignments, and projects. These
classes have the proper associations so that we can access needed
information, as one would expect with these models. I have also set up
the databases for these models with some basic variables (see
migrations). I have also set up the bare bones of the controllers,
however, none of the functions are completed. The needed folders for
the views of these models have also been created, however, there are
not currently any views.
Minor changes include an update to the readme, as well as allowing for
the ‘/‘ route, allowing people to not need to add ‘/snapuser’ at the
end of the url to reach a webpage.